Move current view to legacy folder
